+++ R/svnHelpers.R
@@ -1,10 +1,10 @@
 #' @keywords internal
-svnCommand <- function(.command, .file = NULL, .flags = NULL, .quiet = TRUE) {
+svnCommand <- function(.command, .file = NULL, .flags = NULL, .quiet = TRUE, .xml = TRUE) {
   
   command_run <- paste("svn",
                        .command,
                        .flags,
-                       "--xml",
+                       ifelse(.xml, "--xml", ""),
                        .file,
                        ifelse(.quiet, "2>/dev/null", ""),
                        sep = " ")
@@ -13,6 +13,10 @@ svnCommand <- function(.command, .file = NULL, .flags = NULL, .quiet = TRUE) {
   
   if (!is.null(attr(temp_loc, "status"))) {
     stop("svn command failed")
+  }
+  
+  if (!.xml) {
+    return(invisible(NULL))
   }
   
   parsed_results <- XML::xmlParse(temp_loc)